Move tabs to its oven file
