+++ packages/extension/src/view/devtools/components/sidebar/menuItem.tsx
@@ -16,94 +16,29 @@
 /**
  * External dependencies.
  */
-import React, { type ComponentType, type SVGProps } from 'react';
+import React from 'react';
 
 /**
  * Internal dependencies.
  */
-import {
-  SiteBoundariesIcon,
-  SiteBoundariesIconWhite,
-  ChipsIcon,
-  ChipsIconWhite,
-  RelatedWebsiteSetsIcon,
-  RelatedWebsiteSetsIconWhite,
-  PrivateAdvertisingIcon,
-  PrivateAdvertisingIconWhite,
-  AntiCovertTrackingIcon,
-  AntiCovertTrackingIconWhite,
-  AttributionIcon,
-  AttributionIconWhite,
-  BounceTrackingIcon,
-  BounceTrackingIconWhite,
-  FingerPrintingIcon,
-  FingerPrintingIconWhite,
-  TopicsIcon,
-  TopicsIconWhite,
-} from '../../../../icons';
+import TABS from '../../tabs';
 
 interface MenuItemProps {
   handleClick: () => void;
-  name: string;
+  tabName: string;
+  tabId: string;
   isActive: boolean;
 }
 
-interface Icons {
-  [name: string]: {
-    default: ComponentType<SVGProps<SVGSVGElement>>;
-    selected: ComponentType<SVGProps<SVGSVGElement>>;
-  };
-}
-
-const MenuItem = ({ handleClick, name, isActive }: MenuItemProps) => {
-  const icons: Icons = {
-    'Site Boundaries': {
-      default: SiteBoundariesIcon,
-      selected: SiteBoundariesIconWhite,
-    },
-    CHIPS: {
-      default: ChipsIcon,
-      selected: ChipsIconWhite,
-    },
-    'Related Website Sets': {
-      default: RelatedWebsiteSetsIcon,
-      selected: RelatedWebsiteSetsIconWhite,
-    },
-    'Private Advertising': {
-      default: PrivateAdvertisingIcon,
-      selected: PrivateAdvertisingIconWhite,
-    },
-    Topics: {
-      default: TopicsIcon,
-      selected: TopicsIconWhite,
-    },
-    Attribution: {
-      default: AttributionIcon,
-      selected: AttributionIconWhite,
-    },
-    'Anti-Covert tracking': {
-      default: AntiCovertTrackingIcon,
-      selected: AntiCovertTrackingIconWhite,
-    },
-    'Bounce Tracking': {
-      default: BounceTrackingIcon,
-      selected: BounceTrackingIconWhite,
-    },
-    Fingerprinting: {
-      default: FingerPrintingIcon,
-      selected: FingerPrintingIconWhite,
-    },
-  };
-
-  const Icon = isActive ? icons[name].selected : icons[name].default;
+const MenuItem = ({ handleClick, tabName, tabId, isActive }: MenuItemProps) => {
+  const selectedTab = TABS.find((tab) => tab.id === tabId)?.icons;
+  const Icon = isActive ? selectedTab?.selected : selectedTab?.default;
 
   return (
     <div className="flex w-full items-center pl-6 py-0.5" onClick={handleClick}>
-      <div className="h-3">
-        <Icon />
-      </div>
+      <div className="h-3">{Icon && <Icon />}</div>
       <div className="block">
-        <span className="pl-2.5">{name}</span>
+        <span className="pl-2.5">{tabName}</span>
       </div>
     </div>
   );
